refactor(swap): single computeFiatLossPercent shared by all callers
usePriceDifference (visual colour tier) and getFiatLossWarning (confirm
popup) used to compute the user's USD-value loss separately, with
slightly different precision (.toFixed vs .toExact) and slightly
different bail-out conditions. By construction they could disagree at
the boundary.

Move the calculation into computeFiatLossPercent in the shared fiatLoss
constants module. Both callers now invoke the same function — same
precision, same bail-outs, same result. The colour tier and the popup
trigger are now mathematically guaranteed to fire on the same value.

Side benefit: usePriceDifference now falls back to the pool-curve
impact when USD values are zero/non-finite (previously it produced
NaN and silently dropped the warning).
