+++ packages/uniswap/src/features/transactions/swap/hooks/usePriceDifference.ts
@@ -1,12 +1,12 @@
 import { useMemo } from 'react'
 import { ColorTokens } from 'ui/src'
 import {
+  computeFiatLossPercent,
   FIAT_LOSS_CRITICAL_PERCENT,
   FIAT_LOSS_WARN_PERCENT,
 } from 'uniswap/src/features/transactions/swap/constants/fiatLoss'
 import { DerivedSwapInfo } from 'uniswap/src/features/transactions/swap/types/derivedSwapInfo'
 import { getPriceImpact } from 'uniswap/src/features/transactions/swap/utils/getPriceImpact'
-import { CurrencyField } from 'uniswap/src/types/currency'
 
 export type UsePriceDifferenceReturnType =
   | {
@@ -21,36 +21,34 @@ export type UsePriceDifferenceReturnType =
     }
 
 export function usePriceDifference(derivedSwapInfo?: DerivedSwapInfo): UsePriceDifferenceReturnType {
-  const inputPrice = derivedSwapInfo?.currencyAmountsUSDValue[CurrencyField.INPUT]?.toFixed()
-  const outputPrice = derivedSwapInfo?.currencyAmountsUSDValue[CurrencyField.OUTPUT]?.toFixed()
-
   return useMemo(() => {
     if (!derivedSwapInfo) {
       return { showPriceDifferenceWarning: false }
     }
 
+    // Convention: priceDifferencePercentage is negative when the user is
+    // losing value. computeFiatLossPercent returns positive-for-loss, so we
+    // negate it. Falls back to the pool-curve impact when USD values are
+    // unavailable.
+    const fiatLoss = computeFiatLossPercent(derivedSwapInfo)
     const priceDifferencePercentage =
-      outputPrice !== undefined && inputPrice !== undefined
-        ? ((+outputPrice - +inputPrice) / +inputPrice) * 100
-        : +(getPriceImpact(derivedSwapInfo)?.toFixed() || 0)
+      fiatLoss !== undefined ? -fiatLoss : +(getPriceImpact(derivedSwapInfo)?.toFixed() || 0)
 
-    // Handle NaN cases - return early without warning if calculation is invalid
     if (isNaN(priceDifferencePercentage)) {
       return { showPriceDifferenceWarning: false }
     }
 
-    const showPriceDifferenceWarning =
-      !!priceDifferencePercentage && priceDifferencePercentage * -1 > FIAT_LOSS_WARN_PERCENT
+    const lossPercent = -priceDifferencePercentage
+    const showPriceDifferenceWarning = !!priceDifferencePercentage && lossPercent > FIAT_LOSS_WARN_PERCENT
 
     if (showPriceDifferenceWarning) {
       return {
         priceDifferencePercentage,
         showPriceDifferenceWarning,
-        priceDifferenceColor:
-          priceDifferencePercentage * -1 > FIAT_LOSS_CRITICAL_PERCENT ? '$statusCritical' : '$statusWarning',
+        priceDifferenceColor: lossPercent > FIAT_LOSS_CRITICAL_PERCENT ? '$statusCritical' : '$statusWarning',
       }
     }
 
     return { priceDifferencePercentage, showPriceDifferenceWarning }
-  }, [derivedSwapInfo, inputPrice, outputPrice])
+  }, [derivedSwapInfo])
 }
